+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/type/kotypeparameter/KoTypeParameterDeclarationForKoLocationProviderTest.kt
@@ -7,10 +7,10 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments.arguments
 import org.junit.jupiter.params.provider.MethodSource
 
-class KoTypeParameterDeclarationForKoNameProviderTest {
+class KoTypeParameterDeclarationForKoLocationProviderTest {
     @ParameterizedTest
     @MethodSource("provideValues")
-    fun `type-name`(
+    fun `type-location`(
         fileName: String,
         value: String,
     ) {
@@ -23,13 +23,13 @@ class KoTypeParameterDeclarationForKoNameProviderTest {
                 ?.asTypeParameterDeclaration()
 
         // then
-        sut?.name shouldBeEqualTo value
+        sut?.location shouldBeEqualTo "${sut?.path}:$value"
     }
 
     private fun getSnippetFile(fileName: String) =
         TestSnippetProvider.getSnippetKoScope(
-            "core/declaration/type/kotypeparameter/snippet/forkonameprovider/",
-            fileName
+            "core/declaration/type/kotypeparameter/snippet/forkolocationprovider/",
+            fileName,
         )
 
     companion object {
@@ -37,8 +37,8 @@ class KoTypeParameterDeclarationForKoNameProviderTest {
         @JvmStatic
         fun provideValues() =
             listOf(
-                arguments("type-parameter-name", "TestType"),
-                arguments("nullable-type-parameter-name", "TestType"),
+                arguments("type-parameter-location", "1:43"),
+                arguments("nullable-type-parameter-location", "1:43"),
             )
     }
 }
