+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/type/kotypeparameter/KoTypeParameterDeclarationForKoModuleProviderTest.kt
