+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/type/kotypeparameter/KoTypeParameterDeclarationForKoResideInPackageProviderTest.kt
@@ -12,7 +12,7 @@ class KoTypeParameterDeclarationForKoResideInPackageProviderTest {
         // given
         val sut =
             getSnippetFile("type-parameter-reside-in-package")
-.functions()
+                .functions()
                 .returnTypes
                 .firstOrNull()
                 ?.asTypeParameterDeclaration()
