+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/type/kotypeparameter/KoTypeParameterDeclarationForKoTextProviderTest.kt
@@ -15,7 +15,7 @@ class KoTypeParameterDeclarationForKoTextProviderTest {
         // given
         val sut =
             getSnippetFile(fileName)
-.functions()
+                .functions()
                 .returnTypes
                 .firstOrNull()
                 ?.asTypeParameterDeclaration()
