+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/declaration/type/KoTypeParameterDeclaration.kt
@@ -1,7 +1,5 @@
 package com.lemonappdev.konsist.api.declaration.type
 
-import com.lemonappdev.konsist.api.declaration.KoParentDeclaration
-import com.lemonappdev.konsist.api.provider.KoChildProvider
 import com.lemonappdev.konsist.api.provider.KoContainingDeclarationProvider
 import com.lemonappdev.konsist.api.provider.KoContainingFileProvider
 import com.lemonappdev.konsist.api.provider.KoLocationProvider
