+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/declaration/KoFunctionDeclarationCore.kt
@@ -53,7 +53,6 @@ import org.jetbrains.kotlin.psi.KtDeclarationWithBody
 import org.jetbrains.kotlin.psi.KtElement
 import org.jetbrains.kotlin.psi.KtFunction
 import org.jetbrains.kotlin.psi.KtModifierListOwner
-import org.jetbrains.kotlin.psi.KtTypeParameter
 import org.jetbrains.kotlin.psi.KtTypeParameterListOwner
 
 internal class KoFunctionDeclarationCore private constructor(
