+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/provider/KoTypeDeclarationProviderCore.kt
@@ -45,8 +45,8 @@ internal interface KoTypeDeclarationProviderCore :
 
     private fun isExtensionDeclaration(): Boolean =
         ktTypeReference?.isExtensionDeclaration() == true ||
-                ktNameReferenceExpression?.isExtensionDeclaration() == true ||
-                ktTypeProjection?.isExtensionDeclaration() == true
+            ktNameReferenceExpression?.isExtensionDeclaration() == true ||
+            ktTypeProjection?.isExtensionDeclaration() == true
 
     private fun getDeclarationWithFqn(declaration: KoBaseDeclaration): KoBaseDeclaration? =
         when {
@@ -110,12 +110,12 @@ internal interface KoTypeDeclarationProviderCore :
 
     override fun hasDeclarationOf(kClass: KClass<*>): Boolean =
         hasClassDeclarationOf(kClass) ||
-                hasObjectDeclarationOf(kClass) ||
-                hasInterfaceDeclarationOf(kClass) ||
-                hasKotlinTypeDeclarationOf(
-                    kClass,
-                ) ||
-                hasExternalTypeDeclarationOf(kClass)
+            hasObjectDeclarationOf(kClass) ||
+            hasInterfaceDeclarationOf(kClass) ||
+            hasKotlinTypeDeclarationOf(
+                kClass,
+            ) ||
+            hasExternalTypeDeclarationOf(kClass)
 
     override fun hasClassDeclaration(predicate: ((KoClassDeclaration) -> Boolean)?): Boolean =
         when (predicate) {
@@ -123,8 +123,7 @@ internal interface KoTypeDeclarationProviderCore :
             else -> asClassDeclaration()?.let { predicate(it) } ?: false
         }
 
-    override fun hasClassDeclarationOf(kClass: KClass<*>): Boolean =
-        kClass.qualifiedName == asClassDeclaration()?.fullyQualifiedName
+    override fun hasClassDeclarationOf(kClass: KClass<*>): Boolean = kClass.qualifiedName == asClassDeclaration()?.fullyQualifiedName
 
     override fun hasObjectDeclaration(predicate: ((KoObjectDeclaration) -> Boolean)?): Boolean =
         when (predicate) {
@@ -132,8 +131,7 @@ internal interface KoTypeDeclarationProviderCore :
             else -> asObjectDeclaration()?.let { predicate(it) } ?: false
         }
 
-    override fun hasObjectDeclarationOf(kClass: KClass<*>): Boolean =
-        kClass.qualifiedName == asObjectDeclaration()?.fullyQualifiedName
+    override fun hasObjectDeclarationOf(kClass: KClass<*>): Boolean = kClass.qualifiedName == asObjectDeclaration()?.fullyQualifiedName
 
     override fun hasInterfaceDeclaration(predicate: ((KoInterfaceDeclaration) -> Boolean)?): Boolean =
         when (predicate) {
