+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/provider/KoTypeParameterProviderCore.kt
@@ -1,7 +1,5 @@
 package com.lemonappdev.konsist.core.provider
 
-import com.lemonappdev.konsist.api.provider.KoBodyProvider
-import org.jetbrains.kotlin.psi.KtDeclarationWithBody
 import org.jetbrains.kotlin.psi.KtTypeParameter
 import org.jetbrains.kotlin.psi.KtTypeParameterListOwner
