+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/util/TypeUtil.kt
@@ -5,7 +5,6 @@ import com.lemonappdev.konsist.api.declaration.KoBaseDeclaration
 import com.lemonappdev.konsist.api.declaration.KoFileDeclaration
 import com.lemonappdev.konsist.api.declaration.type.KoBaseTypeDeclaration
 import com.lemonappdev.konsist.api.declaration.type.KoTypeDeclaration
-import com.lemonappdev.konsist.api.declaration.type.KoTypeParameterDeclaration
 import com.lemonappdev.konsist.api.provider.KoContainingDeclarationProvider
 import com.lemonappdev.konsist.api.provider.KoDeclarationProvider
 import com.lemonappdev.konsist.api.provider.KoFullyQualifiedNameProvider
@@ -15,7 +14,6 @@ import com.lemonappdev.konsist.core.declaration.type.KoGenericTypeDeclarationCor
 import com.lemonappdev.konsist.core.declaration.type.KoKotlinTypeDeclarationCore
 import com.lemonappdev.konsist.core.declaration.type.KoStarProjectionDeclarationCore
 import com.lemonappdev.konsist.core.declaration.type.KoTypeParameterDeclarationCore
-import com.lemonappdev.konsist.core.ext.castToKoBaseDeclaration
 import com.lemonappdev.konsist.core.model.getClass
 import com.lemonappdev.konsist.core.model.getInterface
 import com.lemonappdev.konsist.core.model.getObject
@@ -131,11 +129,12 @@ object TypeUtil {
                             .getDeclarationFullyQualifiedName(typeText, parentDeclaration)
                     }
 
-        val hasTypeParameterWithTheSameName = (parentDeclaration as? KoTypeParameterProviderCore)
-            ?.typeParameters
-            ?.map { it.text.substringBefore(":") }
-            ?.map { it.trim() }
-            ?.any { it == typeText }
+        val hasTypeParameterWithTheSameName =
+            (parentDeclaration as? KoTypeParameterProviderCore)
+                ?.typeParameters
+                ?.map { it.text.substringBefore(":") }
+                ?.map { it.trim() }
+                ?.any { it == typeText }
 
         return when {
             nestedType is KtTypeProjection -> KoStarProjectionDeclarationCore.getInstance(nestedType, containingFile)
@@ -147,7 +146,7 @@ object TypeUtil {
                     KoKotlinTypeDeclarationCore.getInstance(nestedType, parentDeclaration)
                 } else if (hasTypeParameterWithTheSameName == true) {
                     KoTypeParameterDeclarationCore.getInstance(nestedType, containingFile)
-                }else {
+                } else {
                     getClass(typeText, fullyQualifiedName, false, containingFile)
                         ?: getInterface(typeText, fullyQualifiedName, false, containingFile)
                         ?: getObject(typeText, fullyQualifiedName, false, containingFile)
